+++ flake.nix
@@ -21,7 +21,7 @@
         {
           default = pkgs.rustPlatform.buildRustPackage {
             pname = "bashlume";
-            version = "0.1.1";
+            version = "0.1.2";
             src = self;
             cargoLock.lockFile = ./Cargo.lock;